FROMLIST: power: supply: qcom_smbx: bump up the max current
I set a super conservative current limit since we lack many of the
safety features (thermal monitoring, etc) that these drivers really
need. However now we have a better understanding of the hardware, it's
fine to bump this limit up a bit, devices can additionally set the max
current via devicetree instead.

Since this is common to smb2 and smb5, move this write out of the init
sequence and into probe proper.

Change-Id: I910d8c06c7304e6d590b6e7c8fa8afb5aff043c1
Signed-off-by: Casey Connolly <casey.connolly@linaro.org>