arm64: dts: qcom: milos-fairphone-fp6: Add dummy panel for simpledrm dimensions
Getting this upstream is a bit complicated with writing bindings for the
panel, which will later be modelled properly as part of MDSS bringup.

Signed-off-by: Luca Weiss <luca.weiss@fairphone.com>

fp6 move dummy panel

This is confusing fw_devlink or something and causing EPROBE_DEFER which
will make the kernel disable clocks for simplefb before it's probed,
which leads to weird issues incl. UFS being slow